feat(stella-tui): Esc steers the running turn instead of restarting it
**The most consequential defect in the report, and the least visible.**

Esc cancelled the in-flight turn unconditionally: the driver truncated
it out of the conversation and auto-dispatched the backlog as a NEW
turn. On the raw step loop that is a defensible "interrupt current, run
next". On the staged pipeline — the deck's default — it is not an
interrupt at all, it is a restart. A new turn re-enters at triage and
context recall, so the plan, the recalled frames, the approved scope
and every file the turn had read are derived again from nothing.

A user who typed one correcting sentence and pressed Esc therefore paid
for the whole turn a second time to deliver it, and watched the deck
drop back to `stage context recall · turn $0.0000` with an approved
six-step plan still reading `0/6`. The sentence did not even arrive: it
stayed in the composer while the turn it was written for died.

Esc now delivers. With anything to say — a draft, a queue, or both —
every waiting text goes into the running turn at its next step boundary,
in order, and the turn keeps going. With nothing to say it is still a
plain interrupt, because with no guidance to deliver that is the only
thing the keystroke can mean, and the double-Esc full stop is untouched.

The whole queue travels in ONE `WorkspaceInput::Steer` rather than N
enqueues, because it must be drained as one act: sent separately, a
second Esc landing between two of them would steer half the backlog and
cancel the turn holding the rest. It also lets the driver take the
settling decision once for all of them — a turn that finished between
the keystroke and the driver reading it routes its texts to the next
turn rather than dropping them.

A steer aimed at a worker lane goes to the backlog instead: a worker has
a steering tap but the driver holds no handle to it (#2896). Weaker than
the lead gets, and stated rather than silently discarded.

Three god files were in the way and none of them was grown: the Esc
decision moved to `deck_ui/steer.rs`, the driver's delivery to
`command_deck/steer.rs`, the reasoning to the doc-only
`envelope/steering.rs`, and `run_lead_pipeline_turn` — a whole execution
mode — out of `command_deck.rs` into `command_deck/pipeline_turn.rs`.

Witnesses: `esc_with_a_typed_draft_steers_the_running_turn_instead_of_cancelling_it`,
`esc_delivers_the_whole_prompt_queue_in_order`,
`esc_with_nothing_to_say_is_still_a_plain_interrupt`.

Behaviour change, named for review: `a_typed_draft_survives_both_esc_forms`
is replaced by the first witness above — its contract (the draft is
preserved and the turn is cancelled) is the defect.
