+++ crates/stella-tui/src/deck_render.rs
@@ -1287,7 +1287,8 @@ fn render_composer_footer(
         Span::styled(" queue (never blocks)", dim),
     ];
     // `>` steers the running turn — the highest-value affordance on this row,
-    // and until now advertised nowhere at all. Shown only while the focused
+    // and until now advertised nowhere at all (`esc` is the shorter reach for
+    // the same act). Shown only while the focused
     // agent is actually running, because that is the only time it does
     // anything: a `>` typed at an idle agent is just the next prompt. It also
     // takes priority over `!`/`/` in the width budget below, since it is the
@@ -1298,7 +1299,7 @@ fn render_composer_footer(
         .is_some_and(|a| a.status == crate::AgentStatus::Running)
     {
         left.push(Span::styled("  ·  ", sep));
-        left.push(Span::styled(">", key));
+        left.push(Span::styled("esc / >", key));
         left.push(Span::styled(" steer this turn", dim));
     }
     let extras = [("!", " shell"), ("/", " commands")];
@@ -1455,10 +1456,7 @@ const GLOBAL_SHORTCUTS: &[(&str, &str)] = &[
         ">text",
         "steer the running turn — lands at the next step boundary",
     ),
-    (
-        "esc",
-        "soft-stop at the next step boundary — completed work kept",
-    ),
+    ("esc", "steer: your draft + queue go INTO the running turn"),
     (
         "esc esc",
         "cancel NOW & hold — nothing runs until your next prompt",
